Remove AGIX, CERRA, COPI, SOC, RJV, rsERG
We're approaching the limit of publication at ~45. We remove
six here, leaving 40. If we expand this list we should keep in
mind we will need to queue publication slightly differently in
the validator and publisher to accomodate more, e.g. split
publication into two to double the feeds. We currently have ~25 on
mainnet.